refactor: enhance reusability with props and name convention
