+++ apps/academy-wep/app/page.tsx
@@ -1,5 +1,7 @@
 'use client'
 
+import CredentialCard from '@/components/credential-card'
+import { HeroSection } from '@/components/hero-section'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { motion } from 'framer-motion'
@@ -134,6 +136,8 @@ export default function Home() {
 					))}
 				</motion.div>
 			</div>
+			<HeroSection />
+			<CredentialCard />
 		</main>
 	)
 }
